hopefully fixes deployment/build related routing issues
