+++ src/App.tsx
@@ -1,15 +1,52 @@
-import { Flex } from '@chakra-ui/layout'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
-import Header from 'components/Header'
-import Page from 'components/Page'
+import theme from 'theme'
+
+import { ChakraProvider, ColorModeScript } from '@chakra-ui/react'
+import { Config, DAppProvider, Mainnet } from '@usedapp/core'
+
+import DPI from 'components/views/DPI'
+import Dashboard from 'components/views/Homepage'
+import LiquidityMining from 'components/views/LiquidityMining'
+import Products from 'components/views/Products'
+import { MarketDataProvider } from 'contexts/MarketData/MarketDataProvider'
+import SetComponentsProvider from 'contexts/SetComponents/SetComponentsProvider'
 
 import './App.css'
+import './index.css'
+
+const config: Config = {
+  readOnlyChainId: Mainnet.chainId,
+  readOnlyUrls: {
+    [Mainnet.chainId]: process.env.REACT_APP_MAINNET_INFURA_API ?? '',
+  },
+}
+
+const Providers = (props: { children: any }) => {
+  return (
+    <ChakraProvider theme={theme}>
+      <DAppProvider config={config}>
+        <MarketDataProvider>
+          <SetComponentsProvider>{props.children}</SetComponentsProvider>
+        </MarketDataProvider>
+      </DAppProvider>
+    </ChakraProvider>
+  )
+}
 
 const App = () => {
   return (
-    <Page>
-      <div>Test Test Test</div>
-    </Page>
+    <Providers>
+      <ColorModeScript initialColorMode={theme.config.initialColorMode} />
+      <BrowserRouter>
+        <Routes>
+          <Route path='/' element={<Dashboard />} />
+          <Route path='/lm' element={<LiquidityMining />} />
+          <Route path='/dpi' element={<DPI />} />
+          <Route path='/products' element={<Products />} />
+        </Routes>
+      </BrowserRouter>
+    </Providers>
   )
 }
